[FRSKYOSD] Add support for increasing the default data rate
Since v1, FrSkyOSD allows changing the baudrate to higher values but
we were ignoring that. Now if the user has configured the respective
port to a higher baudrate, we will switch to it at runtime. This also
requires adding support to probe at different baud rates on startup
since the FC might have been rebooted without rebooting the OSD.